Add Quditto to README.md
The Quditto emulation platform has been added. It is an open-source project that allows the automatic deployment of emulated QKD networks over any Python-capable hardware. The nodes of the emulated network implement the ETSI 014 API, making the interaction between users and said nodes the same as between users and real QKD devices.