Update information for Google's qsim (#148)
* Make a minor correction to the capitalization of Google's qsim

The official spelling of [qsim](https://github.com/quantumlib/qsim) is in fact `qsim`, i.e., all lower case.

* Update description of qsim

Although the original description of qsim was technically correct, I suspect that many people will not have been familiar with the terminology, and will recognize the new description more easily.

* Add an entry for qsim in the Python section

qsim has both C++ and Python interfaces, so it makes sense to put it in both places. It will make it easier for people to find.