Update Qiskit details in full-stack libraries
This commit updates the entry for Qiskit in full-stack libraries to reflect
the current state of the project. There are 3 changes made,
first the URL for the website is updated to it it's current location (the
previous qiskit.org link redirects to the updated url). Second, the
description of the library is updated to use the current description
from the github repository. Finally C is added as a language for
using Qiskit. Qiskit has had a C API since the 2.0 release (see:
https://quantum.cloud.ibm.com/docs/en/api/qiskit-c for the api docs)
and can be used from both Python and C (either together or
completely independently).